Implement integration test
We use the techbench package to run a complete integration test even
though we provide a package - not a complete application.

This test ensures that our service, migrations and routes are
provided as intended.

This supersedes the service provider test which had to jump through
hoops to achieve the same thing.